added messages, moved CoreMessages to messages package
