+++ core/src/main/kotlin/com/github/spigotbasics/core/messages/CoreMessages.kt
@@ -1,7 +1,6 @@
-package com.github.spigotbasics.core.config
+package com.github.spigotbasics.core.messages
 
-import com.github.spigotbasics.core.messages.Message
-import com.github.spigotbasics.core.messages.MessageFactory
+import com.github.spigotbasics.core.config.SavedConfig
 import org.bukkit.permissions.Permission
 import java.io.File
 
@@ -17,6 +16,14 @@ class CoreMessages(file: File, messageFactory: MessageFactory): SavedConfig(file
 
     fun unknownOption(option: String): Message {
         return getMessage("unknown-option").tagUnparsed("option", option)
+    }
+
+    fun invalidArgument(argument: String): Message {
+        return getMessage("invalid-argument").tagUnparsed("argument", argument)
+    }
+
+    fun playerNotFound(name: String): Message {
+        return getMessage("player-not-found").tagUnparsed("argument", name)
     }
 
     val commandNotFromConsole: Message
