+++ core/src/main/resources/messages.yml
@@ -7,5 +7,9 @@ no-permission: "<hover:show_text:'<gray>Permission missing:</gray> <permission>'
 # Tags: <option>
 unknown-option: "<red>Unknown option: </red><color:#FF8888><option></color>"
 
+# Tags: <argument>
+invalid-argument: "<red>Invalid argument: </red><color:#FF8888><argument></color>"
+player-not-found: "<red>Player not found: </red><color:#FF8888><argument></color>"
+
 command-not-from-console: "<red>This command can't be run from console!</red>"
 must-specify-player-from-console: "<red>You must specify a player when running this command from console!</red>"
