+++ plugin/src/main/kotlin/com/github/spigotbasics/plugin/BasicsPluginImpl.kt
@@ -4,7 +4,7 @@ import com.github.spigotbasics.core.BasicsPlugin
 import com.github.spigotbasics.core.Constants
 import com.github.spigotbasics.core.MinecraftVersion
 import com.github.spigotbasics.core.config.CoreConfigManager
-import com.github.spigotbasics.core.config.CoreMessages
+import com.github.spigotbasics.core.messages.CoreMessages
 import com.github.spigotbasics.core.logger.BasicsLoggerFactory
 import com.github.spigotbasics.core.messages.AudienceProvider
 import com.github.spigotbasics.core.messages.MessageFactory
